requirements.txt
Django==2.2.4
django-model-utils==3.2.0
django-multi-token==0.1.2
django-ordered-model==3.2.0
django-silk==3.0.2
djangorestframework==3.10.2
factory-boy==2.12.0
psycopg2==2.8.3
python-decouple==3.1
